Separate test fuction display
