+++ extension/src/extension.ts
@@ -80,40 +80,3 @@ export function deactivate(): Thenable<void> | undefined {
 	}
 	return client.stop();
 }
-
-export class AutoLispDocumentSymbolProvider implements vscode.DocumentSymbolProvider {
-
-    public provideDocumentSymbols(
-        document: vscode.TextDocument,
-        token: vscode.CancellationToken
-    ): Promise<vscode.DocumentSymbol[]> {
-        
-        return new Promise((resolve, reject) => {
-            const symbols: vscode.DocumentSymbol[] = [];
-            
-            // Regex to find AutoLISP function definitions, e.g., (defun C:MYCOMMAND ... )
-            // This regex captures the function name.
-            const functionRegex = /^\s*\(\s*defun\s+([a-zA-Z0-9_:-]+)/gim;
-
-            const text = document.getText();
-            let match;
-            while ((match = functionRegex.exec(text)) !== null) {
-                const functionName = match[1];
-                const position = document.positionAt(match.index);
-                const range = new vscode.Range(position, position); // Simple range
-
-                const symbol = new vscode.DocumentSymbol(
-                    functionName,
-                    'Function', // Detail text shown in the outline
-                    vscode.SymbolKind.Function,
-                    document.lineAt(position.line).range, // The range of the entire line
-                    document.lineAt(position.line).range // The range to select when clicking
-                );
-
-                symbols.push(symbol);
-            }
-
-            resolve(symbols);
-        });
-    }
-}
